#50 Moved Viewport, Window and World to new display.h file.
